[refactor] #65 Integrated logic for the same endpoint in EventDisplayController
